Fixed Failing Tests

+++ tests/resources/functional/snowflake/ddl/test_create_ddl_identity.sql
@@ -11,10 +11,10 @@ CREATE OR REPLACE TABLE sales_data (
 
 -- databricks sql:
 CREATE OR REPLACE TABLE sales_data (
-    sale_id INT GENERATED ALWAYS AS IDENTITY,
-    product_id INT,
-    quantity INT,
+    sale_id  DECIMAL(38, 0) GENERATED ALWAYS AS IDENTITY,
+    product_id  DECIMAL(38, 0),
+    quantity  DECIMAL(38, 0),
     sale_amount DECIMAL(10, 2),
     sale_date DATE,
-    customer_id INT
+    customer_id  DECIMAL(38, 0)
 )
